requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
langchain>=0.0.335
langchain-openai>=0.0.2
langgraph>=0.0.15
chromadb>=0.4.18
crawl4ai>=0.1.0
openai>=1.3.0
python-dotenv>=1.0.0
httpx>=0.25.0
langchain-community>=0.3.0
sentence-transformers
langchain-huggingface
playwright
langchain_chroma